// common/enums/tables.ts
export enum Tables {
  PROFILE = 'Profiles',
  USER = 'Users',
  CATEGORY = 'Categories',
  ORDER = 'Orders',
  FEEDBACK = 'Feedbacks',
  PROFILE_TRANSLATION = 'ProfileTranslations',
  CATEGORY_TRANSLATION = 'CategoryTranslations',
  PROFILE_RATING = 'ProfileRatings',
  EXPORT_DATA = 'ExportDatas',
  NOTIFICATION = 'Notifications',
  CATEGORY_TAGS = 'CategoryTags',
  CHAT = 'Chats',
}